// backend/src/types.ts
export enum LayerParametersEnum {
  priority = 'priority',
  fill = 'fill',
  x = 'x',
  y = 'y',
  opacity = 'opacity',
  shape = 'shape',
  radius = 'radius',
  width = 'width',
  height = 'height',
  points = 'points',
  length = 'length',
  objects = 'objects',
}

type LayerParameters = {
  priority?: number
  fill?: string
  x?: number
  y?: number
  opacity?: number
  shape?: string
  radius?: number
  width?: number
  height?: number
  points?: Array<[number, number]>
  length?: number
  objects?: Identifier[]
}

export enum ObjectParametersEnum {
  priority = 'priority',
  fill = 'fill',
  x = 'x',
  y = 'y',
  opacity = 'opacity',
  shape = 'shape',
  radius = 'radius',
  width = 'width',
  height = 'height',
  points = 'points',
  length = 'length',
}
// ...
export enum AnimationParametersEnum {
  duration = 'duration',
  delay = 'delay',
  velocity = 'velocity',
  direction = 'direction',
  angularVelocity = 'angularVelocity',
  scaleFactor = 'scaleFactor',
  fill = 'fill',
  opacity = 'opacity',
}
// ...
export type Parameter = { parameter: string; value: any }